pyproject.toml
[project]
name = "vvalkman"
version = "0.1.0"
description = "Open Round chat model demo"
requires-python = ">=3.12"

dependencies = [
    "requests",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "langchain-google-genai>=2.1.12",
    "openai",
    "google>=3.0.0",
    "google-genai>=1.52.0",
    "pytest",
    "python-dotenv",
    "pydantic>=2.11.7",
    "fastapi",
    "uvicorn",
    "chromadb",
    "sentence-transformers",
    "langchain-chroma",
    "langchain-huggingface",
    "huggingface-hub[cli]>=0.36.0",
    "langgraph",
    "pyspark",
    "wikipedia",
    "gitpython",
    "sqlalchemy>=2.0.43",
    "mcp>=1.25.0",
    "asyncpg>=0.31.0",
    "greenlet>=3.2.4",
    "aiosqlite>=0.19.0",
    "langchain-deepseek>=0.1.4",
    "langchain-ollama>=0.3.10",
    "langchain-postgres>=0.0.16",
    "langchain-core>=0.3.76",
    "langfuse>=2.60.10,<3.0.0",
    "mlx-lm>=0.29.1",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["test"]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true 

[dependency-groups]
dev = [
    "black>=25.12.0",
    "ipykernel>=7.1.0",
    "mypy>=1.19.1",
    "ruff>=0.14.10",
]



